refactor: organize folder structure to remove `v2` subfolder
