+++ src/mock-data/posts.js
@@ -1,5 +1,5 @@
 import posts from './posts.json'
-import cloneUtils from '../../utils/shallow-clone-entity'
+import cloneUtils from '../utils/shallow-clone-entity'
 
 // lodash
 import find from 'lodash/find'
